# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test Vehicle API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/users/{id}/trips": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of Trip by User",
                "operationId": "04438097f4cc5d65d7eb2e90526b1c9d",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/User--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trips"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/users/{user_id}/trips/{trip_id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get Trip by User and id",
                "operationId": "8acd8ff53cd2d9b8dc2aed86a005d05b",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/User--id"
                    },
                    {
                        "$ref": "#/components/parameters/Trip--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Trip"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Trip or User not found"
                    }
                }
            }
        },
        "/api/vehicles": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Get list of Vehicle",
                "operationId": "727634867f59796290a2fff276979198",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vehicles"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Insert a new Vehicle",
                "operationId": "a88e0bca4193c91e713d0441659d199c",
                "requestBody": {
                    "description": "Vehicle to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Vehicle"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Vehicle"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/api/vehicles/{id}": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Get Vehicle by id",
                "operationId": "0185374b4b707b85ca8a97a6f5f366e1",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Vehicle--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Vehicle"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehicle not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Updates a Vehicle",
                "operationId": "c4d991b5c398a7ccd4e73f9a81c355d5",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Vehicle--id"
                    }
                ],
                "requestBody": {
                    "description": "Vehicle to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Vehicle"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Vehicle"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehicle not found"
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Delete a Vehicle",
                "operationId": "3442f99c22f47affb335443ef012e712",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Vehicle--id"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Vehicle not found"
                    }
                }
            }
        },
        "/api/vehicles/{id}/trips": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Get list of Trip by Vehicle",
                "operationId": "cc21b998b6f826cb14cb838ea5a71f5c",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Vehicle--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trips"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehicle not found"
                    }
                }
            }
        },
        "/api/vehicles/{vehicle_id}/trips/{trip_id}": {
            "get": {
                "tags": [
                    "Vehicles"
                ],
                "summary": "Get Trip by Vehicle and id",
                "operationId": "ba511e60b22b17f7c837cb996f350074",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Vehicle--id"
                    },
                    {
                        "$ref": "#/components/parameters/Trip--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Trip"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Trip or Vehicle not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Trip": {
                "required": [],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "comment": {
                        "type": "string",
                        "example": "Grocery shopping trip"
                    },
                    "user_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "vehicle_id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "start_at": {
                        "type": "string",
                        "format": "datetime",
                        "readOnly": true,
                        "example": "2022-01-02 12:13:14"
                    },
                    "finish_at": {
                        "type": "string",
                        "format": "datetime",
                        "readOnly": true,
                        "example": "2022-01-02 13:00:00",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "readOnly": true,
                        "example": "2022-01-02 12:13:14"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime",
                        "readOnly": true,
                        "example": "2022-01-02 12:13:14"
                    }
                },
                "type": "object"
            },
            "Trips": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Trip"
                        }
                    }
                },
                "type": "object"
            },
...
            "Vehicle": {
                "required": [],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "model": {
                        "type": "string",
                        "example": "Volvo"
                    },
                    "number": {
                        "type": "string",
                        "example": "ABC-123"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "datetime",
                        "readOnly": true,
                        "example": "2022-01-02 12:13:14"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "datetime",
                        "readOnly": true,
                        "example": "2022-01-02 12:13:14"
                    }
                },
                "type": "object"
            },
            "Vehicles": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Vehicle"
                        }
                    }
                },
                "type": "object"
            }
        },
        "parameters": {
            "Trip--id": {
                "name": "trip_id",
                "in": "path",
                "description": "Id of Trip",
                "required": true,
                "schema": {
                    "type": "integer"
                },
                "example": 1
            },
            "User--id": {
                "name": "user_id",
                "in": "path",
                "description": "Id of User",
                "required": true,
                "schema": {
                    "type": "integer"
                },
                "example": 1
            },
            "Vehicle--id": {
                "name": "vehicle_id",
                "in": "path",
                "description": "Id of Vehicle",
                "required": true,
                "schema": {
                    "type": "integer"
                },
                "example": 1
            }
        }
    },
...
}